feat(suite-native): old onboarding module renamed to module-onboarding-legacy
