+++ suite-native/app/src/navigation/RootStackNavigator.tsx
@@ -16,6 +16,7 @@ import {
 import { selectIsOnboardingFinished } from '@suite-native/settings';
 import { DevUtilsStackNavigator } from '@suite-native/module-dev-utils';
 import { TransactionDetailScreen } from '@suite-native/transactions';
+import { OnboardingStackNavigator as LegacyOnboardingStackNavigator } from '@suite-native/module-onboarding-legacy';
 import { OnboardingStackNavigator } from '@suite-native/module-onboarding';
 import { ReceiveModalScreen } from '@suite-native/receive';
 import { AuthorizeDeviceStackNavigator } from '@suite-native/module-authorize-device';
@@ -44,7 +45,7 @@ export const RootStackNavigator = () => {
             return RootStackRoutes.AppTabs;
         }
 
-        return RootStackRoutes.Onboarding;
+        return RootStackRoutes.LegacyOnboarding;
     };
 
     return (
@@ -52,6 +53,10 @@ export const RootStackNavigator = () => {
             initialRouteName={getInitialRouteName()}
             screenOptions={stackNavigationOptionsConfig}
         >
+            <RootStack.Screen
+                name={RootStackRoutes.LegacyOnboarding}
+                component={LegacyOnboardingStackNavigator}
+            />
             <RootStack.Screen
                 name={RootStackRoutes.Onboarding}
                 component={OnboardingStackNavigator}
