+++ suite-native/app/tsconfig.json
@@ -1,6 +1,8 @@
 {
     "extends": "../../tsconfig.base.json",
-    "compilerOptions": { "outDir": "./libDev" },
+    "compilerOptions": {
+        "outDir": "./libDev"
+    },
     "references": [
         {
             "path": "../../suite-common/analytics"
@@ -29,51 +31,126 @@
         {
             "path": "../../suite-common/wallet-types"
         },
-        { "path": "../accounts" },
-        { "path": "../alerts" },
-        { "path": "../analytics" },
-        { "path": "../atoms" },
-        { "path": "../biometrics" },
-        { "path": "../coin-enabling" },
-        { "path": "../config" },
-        { "path": "../connection-status" },
-        { "path": "../device" },
-        { "path": "../device-authorization" },
-        { "path": "../discovery" },
-        { "path": "../icons" },
-        { "path": "../intl" },
-        { "path": "../message-system" },
-        { "path": "../module-accounts-import" },
+        {
+            "path": "../accounts"
+        },
+        {
+            "path": "../alerts"
+        },
+        {
+            "path": "../analytics"
+        },
+        {
+            "path": "../atoms"
+        },
+        {
+            "path": "../biometrics"
+        },
+        {
+            "path": "../coin-enabling"
+        },
+        {
+            "path": "../config"
+        },
+        {
+            "path": "../connection-status"
+        },
+        {
+            "path": "../device"
+        },
+        {
+            "path": "../device-authorization"
+        },
+        {
+            "path": "../discovery"
+        },
+        {
+            "path": "../icons"
+        },
+        {
+            "path": "../intl"
+        },
+        {
+            "path": "../message-system"
+        },
+        {
+            "path": "../module-accounts-import"
+        },
         {
             "path": "../module-accounts-management"
         },
-        { "path": "../module-add-accounts" },
-        { "path": "../module-authorize-device" },
-        { "path": "../module-connect-popup" },
-        { "path": "../module-dev-utils" },
-        { "path": "../module-device-settings" },
-        { "path": "../module-home" },
-        { "path": "../module-onboarding" },
-        { "path": "../module-send" },
-        { "path": "../module-settings" },
+        {
+            "path": "../module-add-accounts"
+        },
+        {
+            "path": "../module-authorize-device"
+        },
+        {
+            "path": "../module-connect-popup"
+        },
+        {
+            "path": "../module-dev-utils"
+        },
+        {
+            "path": "../module-device-settings"
+        },
+        {
+            "path": "../module-home"
+        },
+        {
+            "path": "../module-onboarding"
+        },
+        {
+            "path": "../module-onboarding-legacy"
+        },
+        {
+            "path": "../module-send"
+        },
+        {
+            "path": "../module-settings"
+        },
         {
             "path": "../module-staking-management"
         },
-        { "path": "../navigation" },
-        { "path": "../notifications" },
-        { "path": "../receive" },
-        { "path": "../settings" },
-        { "path": "../state" },
-        { "path": "../storage" },
-        { "path": "../theme" },
-        { "path": "../toasts" },
-        { "path": "../transactions" },
-        { "path": "../../packages/connect" },
+        {
+            "path": "../navigation"
+        },
+        {
+            "path": "../notifications"
+        },
+        {
+            "path": "../receive"
+        },
+        {
+            "path": "../settings"
+        },
+        {
+            "path": "../state"
+        },
+        {
+            "path": "../storage"
+        },
+        {
+            "path": "../theme"
+        },
+        {
+            "path": "../toasts"
+        },
+        {
+            "path": "../transactions"
+        },
+        {
+            "path": "../../packages/connect"
+        },
         {
             "path": "../../packages/react-native-usb"
         },
-        { "path": "../../packages/styles" },
-        { "path": "../../packages/theme" },
+        {
+            "path": "../../packages/styles"
+        },
+        {
+            "path": "../../packages/theme"
+        },
         {
             "path": "../../packages/trezor-user-env-link"
         },
@@ -84,5 +161,8 @@
             "path": "../../packages/connect-mobile"
         }
     ],
-    "include": [".", "**.json"]
+    "include": [
+        ".",
+        "**.json"
+    ]
 }
