+++ suite-native/module-onboarding-legacy/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-native/module-onboarding",
+    "name": "@suite-native/module-onboarding-legacy",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
