+++ suite-native/module-onboarding-legacy/src/assets/darkRectangles.png
