+++ suite-native/module-onboarding-legacy/src/assets/darkRectangles@2x.png
