+++ suite-native/module-onboarding-legacy/src/assets/darkRectangles@3x.png
