+++ suite-native/module-onboarding-legacy/src/assets/rectangles.png
