+++ suite-native/module-onboarding-legacy/src/assets/rectangles@2x.png
