+++ suite-native/module-onboarding-legacy/src/assets/rectangles@3x.png
