+++ suite-native/module-onboarding-legacy/src/components/AnalyticsInfoRow.tsx
