+++ suite-native/module-onboarding-legacy/src/components/CoinsSvg.tsx
