+++ suite-native/module-onboarding-legacy/src/components/GraphSvg.tsx
