+++ suite-native/module-onboarding-legacy/src/components/OnboardingFooter.tsx
