+++ suite-native/module-onboarding-legacy/src/components/OnboardingScreen.tsx
