+++ suite-native/module-onboarding-legacy/src/components/OnboardingScreenHeader.tsx
