+++ suite-native/module-onboarding-legacy/src/index.ts
