+++ suite-native/module-onboarding-legacy/src/navigation/OnboardingStackNavigator.tsx
@@ -1,8 +1,8 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import {
-    OnboardingStackParamList,
-    OnboardingStackRoutes,
+    LegacyOnboardingStackRoutes,
+    LegacyOnboardingStackParamList,
     stackNavigationOptionsConfig,
 } from '@suite-native/navigation';
 
@@ -12,28 +12,31 @@ import { FeatureReceiveScreen } from '../screens/FeatureReceiveScreen';
 import { AnalyticsConsentScreen } from '../screens/AnalyticsConsentScreen';
 import { ConnectTrezorScreen } from '../screens/ConnectTrezorScreen';
 
-export const OnboardingStack = createNativeStackNavigator<OnboardingStackParamList>();
+export const OnboardingStack = createNativeStackNavigator<LegacyOnboardingStackParamList>();
 
 export const OnboardingStackNavigator = () => (
     <OnboardingStack.Navigator
-        initialRouteName={OnboardingStackRoutes.Welcome}
+        initialRouteName={LegacyOnboardingStackRoutes.Welcome}
         screenOptions={stackNavigationOptionsConfig}
     >
-        <OnboardingStack.Screen name={OnboardingStackRoutes.Welcome} component={WelcomeScreen} />
         <OnboardingStack.Screen
-            name={OnboardingStackRoutes.TrackBalances}
+            name={LegacyOnboardingStackRoutes.Welcome}
+            component={WelcomeScreen}
+        />
+        <OnboardingStack.Screen
+            name={LegacyOnboardingStackRoutes.TrackBalances}
             component={TrackBalancesScreen}
         />
         <OnboardingStack.Screen
-            name={OnboardingStackRoutes.AboutReceiveCoinsFeature}
+            name={LegacyOnboardingStackRoutes.AboutReceiveCoinsFeature}
             component={FeatureReceiveScreen}
         />
         <OnboardingStack.Screen
-            name={OnboardingStackRoutes.ConnectTrezor}
+            name={LegacyOnboardingStackRoutes.ConnectTrezor}
             component={ConnectTrezorScreen}
         />
         <OnboardingStack.Screen
-            name={OnboardingStackRoutes.AnalyticsConsent}
+            name={LegacyOnboardingStackRoutes.AnalyticsConsent}
             component={AnalyticsConsentScreen}
         />
     </OnboardingStack.Navigator>
