+++ suite-native/module-onboarding-legacy/tsconfig.json
