+++ suite-native/navigation/src/navigators.ts
@@ -22,6 +22,7 @@ import {
     DeviceStackRoutes,
     DevUtilsStackRoutes,
     HomeStackRoutes,
+    LegacyOnboardingStackRoutes,
     OnboardingStackRoutes,
     ReceiveStackRoutes,
     RootStackRoutes,
@@ -103,12 +104,16 @@ export type AppTabsParamList = {
     [AppTabsRoutes.Settings]: undefined;
 };
 
+export type LegacyOnboardingStackParamList = {
+    [LegacyOnboardingStackRoutes.Welcome]: undefined;
+    [LegacyOnboardingStackRoutes.AboutReceiveCoinsFeature]: undefined;
+    [LegacyOnboardingStackRoutes.TrackBalances]: undefined;
+    [LegacyOnboardingStackRoutes.AnalyticsConsent]: undefined;
+    [LegacyOnboardingStackRoutes.ConnectTrezor]: undefined;
+};
+
 export type OnboardingStackParamList = {
     [OnboardingStackRoutes.Welcome]: undefined;
-    [OnboardingStackRoutes.AboutReceiveCoinsFeature]: undefined;
-    [OnboardingStackRoutes.TrackBalances]: undefined;
-    [OnboardingStackRoutes.AnalyticsConsent]: undefined;
-    [OnboardingStackRoutes.ConnectTrezor]: undefined;
 };
 
 export type AccountsImportStackParamList = {
@@ -190,7 +195,8 @@ export type AuthorizeDeviceStackParamList = {
 
 export type RootStackParamList = {
     [RootStackRoutes.AppTabs]: NavigatorScreenParams<AppTabsParamList>;
-    [RootStackRoutes.Onboarding]: NavigatorScreenParams<AppTabsParamList>;
+    [RootStackRoutes.LegacyOnboarding]: NavigatorScreenParams<LegacyOnboardingStackParamList>;
+    [RootStackRoutes.Onboarding]: NavigatorScreenParams<OnboardingStackParamList>;
     [RootStackRoutes.AuthorizeDeviceStack]: NavigatorScreenParams<AuthorizeDeviceStackParamList>;
     [RootStackRoutes.AccountsImport]: NavigatorScreenParams<AccountsImportStackParamList>;
     [RootStackRoutes.ReceiveModal]: AccountDetailParams;
